requirements.txt
langchain-google-genai
langchain-core
langchain-community
python-dotenv
streamlit